requirements.txt
Django==5.2.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-jazzmin
markdown
Pillow
python-dotenv==1.0.1
django-simple-history